requirements.txt
Pillow>=9.0.0
requests>=2.25.0
Flask==2.3.3
Werkzeug==2.3.7 